improving speedups
