+++ devtools/__init__.py
@@ -1,4 +1,5 @@
 # flake8: noqa
-from .debug import Debug, debug
-from .prettier import pformat
+from .debug import *
+from .prettier import *
+from .timer import *
 from .version import VERSION
